projects/ria/pip-requirements.txt
mammoth
openpyxl
pandas
xls2xlsx
xlsx2html
psycopg2-binary
flatdict